requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
python-dateutil==2.8.2
email-validator==2.1.0
aiofiles==23.2.1
minio==7.2.0 